src/app/requirements.txt
greenlet==3.2.4
numpy==2.3.4
pandas==2.3.3
pyodbc==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.44
tabulate==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
